Add vanilla CSS support to Progress (#525)
