+++ components/demos/Progress/css/index.jsx
@@ -1,0 +1,23 @@
+import React from 'react';
+import * as Progress from '@radix-ui/react-progress';
+import './styles.css';
+
+const ProgressDemo = () => {
+  const [progress, setProgress] = React.useState(13);
+
+  React.useEffect(() => {
+    const timer = setTimeout(() => setProgress(66), 500);
+    return () => clearTimeout(timer);
+  }, []);
+
+  return (
+    <Progress.Root className="ProgressRoot" value={66}>
+      <Progress.Indicator
+        className="ProgressIndicator"
+        style={{ transform: `translateX(-${100 - progress}%)` }}
+      />
+    </Progress.Root>
+  );
+};
+
+export default ProgressDemo;
