+++ components/demos/Progress/stitches/index.jsx
@@ -1,0 +1,41 @@
+import React from 'react';
+import * as Progress from '@radix-ui/react-progress';
+import { styled } from '@stitches/react';
+import { blackA } from '@radix-ui/colors';
+
+const ProgressDemo = () => {
+  const [progress, setProgress] = React.useState(13);
+
+  React.useEffect(() => {
+    const timer = setTimeout(() => setProgress(66), 500);
+    return () => clearTimeout(timer);
+  }, []);
+
+  return (
+    <ProgressRoot value={66}>
+      <ProgressIndicator style={{ transform: `translateX(-${100 - progress}%)` }} />
+    </ProgressRoot>
+  );
+};
+
+const ProgressRoot = styled(Progress.Root, {
+  position: 'relative',
+  overflow: 'hidden',
+  background: blackA.blackA9,
+  borderRadius: '99999px',
+  width: 300,
+  height: 25,
+
+  // Fix overflow clipping in Safari
+  // https://gist.github.com/domske/b66047671c780a238b51c51ffde8d3a0
+  transform: 'translateZ(0)',
+});
+
+const ProgressIndicator = styled(Progress.Indicator, {
+  backgroundColor: 'white',
+  width: '100%',
+  height: '100%',
+  transition: 'transform 660ms cubic-bezier(0.65, 0, 0.35, 1)',
+});
+
+export default ProgressDemo;
