+++ data/primitives/components/progress/1.0.1.mdx
@@ -16,9 +16,7 @@ aria: https://www.w3.org/WAI/ARIA/apg/patterns/meter
   <ProgressDemo />
 </HeroContainer>
 
-```jsx hero template=Progress
-
-```
+<HeroCodeBlock folder="Progress" />
 
 <Highlights
   features={[
